Jose_Sanez_Juggling_V2
This is my progress so far. I had to restart because I messed up the
keying for the right arm of the previous rig and the hip movement was
off. Also, I keep getting an error with the file format. Somehow the
file got changed from .mb to .ma and does not let me save it as a .mb